+++ bee-identity/CHANGELOG.md
@@ -33,4 +33,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Identity` type;
 - `pem_file` module;
+- `migration` module;